move geo validation to shared API, validate areas
