+++ api/src/utils/batch/spatial-validation.ts
@@ -1,6 +1,6 @@
-import { getDBConnection } from '../../database/db';
+import {getDBConnection} from '../../database/db';
 
-import { parse } from 'wkt';
+import {parse} from 'wkt';
 
 export const validateAsWKT = (input: string) => {
   try {
@@ -24,6 +24,23 @@ export const checkWKTInBounds = async (input: string) => {
       values: [input]
     });
     return res.rows[0]['valid'];
+  } finally {
+    connection.release();
+  }
+};
+
+export const computeWKTArea = async (input: string) => {
+  const connection = await getDBConnection();
+
+  if (!connection) {
+    throw new Error('Could not get a DB Connection');
+  }
+  try {
+    const res = await connection.query({
+      text: `SELECT ST_Area(ST_GeomFromText($1)::geography) as area`,
+      values: [input]
+    });
+    return res.rows[0]['area'];
   } finally {
     connection.release();
   }
