+++ sharedAPI/src/validation/herbicideApplicationRates.ts
