+++ sharedAPI/src/validation/herbicideCalculator.ts
